requirements.txt
Django==2.2
graphene-django==2.2.0
django-filter==2.1.0
django-graphql-jwt==0.2.1
django-cors-headers==2.5.2
